Fixed date alignment issue by deleting constraint that was holding it to the bottom of the screen. This was causing the date to not show up in 3.5 inch devices. This was on the FuelDetailViewController. I also fixed several NSString stringWithFormat warnings about using NSNumbers and casted them all to (long). Now the only warnings are coming from the view controllers that I haven't implemented yet.Oh yeah, I fixed the CarTableView so that it shows the background image through the tableviewcells. I am thinking about adding a slight blur so that when the cells cover the background they kind of "pop". Don't know if it will look good. Just something I'm thinking about doing.

+++ MileageTracker/CarDetailTVController.m
@@ -40,17 +40,17 @@
     _carNameTextField.text  = _selectedCar.name;
     _carMakeTextField.text  = _selectedCar.make;
     _carVINTextField.text   = _selectedCar.VIN;
-    NSString *yearString = [NSString stringWithFormat:@"%li", _selectedCar.year];
+    NSString *yearString = [NSString stringWithFormat:@"%li", (long)_selectedCar.year];
     _carYearTextField.text  = yearString;
     if ([self.selectedCar.uuid isEqualToString:[self activeCarID]]) {
         _activeVehicleLabel.text = @"Yes";
     } else {
         _activeVehicleLabel.text = @"No";
     }
-    NSString *milesString = [NSString stringWithFormat:@"%li", _selectedCar.currentMileage];
+    NSString *milesString = [NSString stringWithFormat:@"%li", (long)_selectedCar.currentMileage];
     _carMileageTextField.text = milesString;
     _fuelCardPINTextField.text = _selectedCar.driverID;
-    NSString *oilChangeMiles = [NSString stringWithFormat:@"%li", _selectedCar.oilChangeMiles];
+    NSString *oilChangeMiles = [NSString stringWithFormat:@"%li", (long)_selectedCar.oilChangeMiles];
     _carOilChangeMileageTextField.text = oilChangeMiles;
     PhotoObject *obj = [self.selectedCar.carPhoto firstObject];
